avm1: Moved root_error_handler from avm1 root to runtime module
